Standartize tests
