+++ src/btc/test.rs
@@ -1,6 +1,6 @@
-use ergvein_filters::btc::ErgveinFilter;
-use ergvein_filters::mempool::ErgveinMempoolFilter;
-use ergvein_filters::util::is_script_indexable;
+use crate::btc::ErgveinFilter;
+use crate::test::utils::*;
+use crate::util::is_script_indexable;
 use bitcoin::consensus::deserialize;
 use bitcoin::hashes::hex::FromHex;
 use bitcoin::util::bip158::Error;
@@ -8,8 +8,6 @@ use bitcoin::BlockHash;
 
 use bitcoin::Transaction;
 
-mod util;
-use crate::util::*;
 
 #[test]
 fn block_00000000000000000007fc62780dee62d79ba02e7d325d7503e80c4da8b16b72() {
@@ -54,39 +52,6 @@ fn block_000000000000017c36b1c7c70f467244009c552e1732604a0f779fc6ff2d6112() {
         if is_indexable {
             assert!(
                 filter.match_tx_outputs(&block_hash, tx).unwrap(),
-                "Tx #{} failed",
-                i
-            );
-        }
-    }
-}
-
-#[test]
-fn mempool_test() {
-    let k0 = u64::from_le_bytes(*b"qwertyui");
-    let k1 = u64::from_le_bytes(*b"opasdfgh");
-    let block = load_block("./test/block1");
-    let txmap = make_inputs_map(load_txs("./test/block1-txs"));
-    let mut txs = block.txdata;
-    txs.remove(0); // remove coinbase
-    let txs2 = txs.clone();
-    let filter = ErgveinMempoolFilter::new_script_filter(k0, k1, txs, |o| {
-        if let Some(s) = txmap.get(o) {
-            Ok(s.clone())
-        } else {
-            Err(Error::UtxoMissing(o.clone()))
-        }
-    })
-    .unwrap();
-
-    for (i, tx) in txs2.iter().enumerate() {
-        let is_indexable = tx
-            .output
-            .iter()
-            .any(|o| is_script_indexable(&o.script_pubkey));
-        if is_indexable {
-            assert!(
-                filter.match_tx_outputs(k0, k1, &tx).unwrap(),
                 "Tx #{} failed",
                 i
             );
