+++ src/lib.rs
@@ -1,3 +1,6 @@
 pub mod btc;
 pub mod mempool;
 pub mod util;
+
+#[cfg(test)]
+pub mod test;
