+++ src/mempool.rs
@@ -4,6 +4,9 @@ use bitcoin::{OutPoint, Script, Transaction};
 use std::io;
 use std::io::Cursor;
 
+#[cfg(test)]
+mod test;
+
 /// Golomb encoding parameter as in BIP-158, see also https://gist.github.com/sipa/576d5f09c3b86c3b1b75598d799fc845
 const P: u8 = 19;
 const M: u64 = 784931;
